Formatting and move color picker dialog to its own file
