+++ app/src/main/java/com/example/palletify/ui/preview/PreviewScreen.kt
@@ -4,6 +4,7 @@ package com.example.palletify.ui.preview
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.border
+import androidx.compose.foundation.horizontalScroll
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxWithConstraints
@@ -54,34 +55,19 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import androidx.compose.ui.window.Dialog
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.example.palletify.ColorUtils.hexToComposeColor
-import com.example.palletify.R
-import com.example.palletify.ui.theme.PalletifyTheme
-import kotlin.random.Random
-import android.util.Log
-import androidx.compose.foundation.horizontalScroll
-import com.github.skydoves.colorpicker.compose.HsvColorPicker
-import com.github.skydoves.colorpicker.compose.AlphaSlider
-import com.github.skydoves.colorpicker.compose.BrightnessSlider
-import com.github.skydoves.colorpicker.compose.rememberColorPickerController
-import com.github.skydoves.colorpicker.compose.ColorPickerController
-import com.github.skydoves.colorpicker.compose.AlphaTile
-import com.example.palletify.data.Palette
-import androidx.compose.ui.window.Dialog
 import com.example.palletify.ColorUtils
+import com.example.palletify.ColorUtils.hexToComposeColor
+import com.example.palletify.ColorUtils.toPaletteColorString
+import com.example.palletify.R
 import com.example.palletify.database.PaletteViewModel
+import com.example.palletify.ui.components.ColorPickerDialog
+import com.example.palletify.ui.theme.PalletifyTheme
+import com.github.skydoves.colorpicker.compose.rememberColorPickerController
+import kotlin.random.Random
 
-fun Color.toPaletteColor(): String {
-    // Convert the floating-point values to 0-255 and then to a hex string
-    val redValue = (this.red * 255).toInt()
-    val greenValue = (this.green * 255).toInt()
-    val blueValue = (this.blue * 255).toInt()
-    val alphaValue = (this.alpha * 255).toInt()
-
-    return String.format("#%02X%02X%02X%02X", alphaValue, redValue, greenValue, blueValue)
-}
 
 @Composable
 fun OutlinedRadioButtonWithText(
@@ -249,86 +235,6 @@ fun RadioButtonGroup(
 }
 
 @Composable
-fun ColorPickerDialog(
-    controller: ColorPickerController,
-    onDismissRequest: () -> Unit,
-    initialColor: Color,
-    onColorChanged: (Color) -> Unit,
-    onSave: (Color) -> Unit
-) {
-    Dialog(onDismissRequest = onDismissRequest) {
-        Card {
-            Column(
-                modifier = Modifier.padding(16.dp),
-                verticalArrangement = Arrangement.spacedBy(8.dp)
-            ) {
-                HsvColorPicker(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .heightIn(max = 300.dp)
-                        .padding(10.dp),
-                    controller = controller,
-                    initialColor = initialColor
-                )
-
-                Spacer(modifier = Modifier.height(16.dp))
-
-                AlphaSlider(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(10.dp)
-                        .height(35.dp),
-                    controller = controller,
-                    initialColor = initialColor
-                )
-                Spacer(modifier = Modifier.height(16.dp))
-
-                BrightnessSlider(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(10.dp)
-                        .height(35.dp),
-                    controller = controller,
-                    initialColor = initialColor
-                )
-
-                Spacer(modifier = Modifier.height(16.dp))
-
-                AlphaTile(
-                    modifier = Modifier
-                        .size(80.dp),
-                    controller = controller
-                )
-
-                Spacer(modifier = Modifier.height(16.dp))
-
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth(),
-                    horizontalArrangement = Arrangement.End
-                ) {
-                    Button(
-                        onClick = onDismissRequest,
-                        modifier = Modifier.padding(end = 8.dp)
-                    ) {
-                        Text("Cancel")
-                    }
-                    Button(
-                        onClick = {
-                            val c = controller.selectedColor
-                            onSave(c.value)
-                            onDismissRequest()
-                        }
-                    ) {
-                        Text("Save")
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
 fun ContrastColorPickerDialog(
     palette: List<String>,
     onDismissRequest: () -> Unit,
@@ -343,22 +249,29 @@ fun ContrastColorPickerDialog(
                 verticalArrangement = Arrangement.spacedBy(8.dp)
             ) {
                 Row(
-                    modifier=Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.Start
+                    modifier = Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.Start
                 ) {
-                    Text("Color One: ", modifier = Modifier.padding(8.dp), style=TextStyle(fontSize =24.sp))
-                    ColorSample(color=firstColor, onClick={})
+                    Text(
+                        "Color One: ",
+                        modifier = Modifier.padding(8.dp),
+                        style = TextStyle(fontSize = 24.sp)
+                    )
+                    ColorSample(color = firstColor, onClick = {})
                 }
 
-                Text("Select Contrast Color", style = TextStyle(
-                    fontSize = 24.sp,
-                    textAlign = TextAlign.Center
-                ),)
+                Text(
+                    "Select Contrast Color",
+                    style = TextStyle(
+                        fontSize = 24.sp,
+                        textAlign = TextAlign.Center
+                    ),
+                )
                 Row(
                     modifier = Modifier.horizontalScroll(rememberScrollState()),
                     horizontalArrangement = Arrangement.spacedBy(8.dp)
                 ) {
                     palette.forEach { color ->
-                        ColorSample(color = color, onClick={
+                        ColorSample(color = color, onClick = {
                             val contrastRatio = ColorUtils.contrastRatio(
                                 hexToComposeColor(firstColor),
                                 hexToComposeColor(color)
@@ -375,10 +288,13 @@ fun ContrastColorPickerDialog(
                 Spacer(modifier = Modifier.height(16.dp))
 
                 contrastResult.value?.let {
-                    Text(it, style = TextStyle(
-                        fontSize = 24.sp,
-                        textAlign = TextAlign.Center
-                    ),)
+                    Text(
+                        it,
+                        style = TextStyle(
+                            fontSize = 24.sp,
+                            textAlign = TextAlign.Center
+                        ),
+                    )
                 }
 
                 Spacer(modifier = Modifier.height(16.dp))
@@ -422,7 +338,7 @@ fun PreviewScreen(previewViewModel: PreviewViewModel = viewModel()) {
     val showColorPicker = remember { mutableStateOf(false) }
     val showContrastModal = remember { mutableStateOf(false) }
     val controller = rememberColorPickerController()
-    val palleteViewModel: PaletteViewModel = viewModel()
+    val paletteViewModel: PaletteViewModel = viewModel()
 
     Scaffold(
         topBar = {
@@ -462,15 +378,15 @@ fun PreviewScreen(previewViewModel: PreviewViewModel = viewModel()) {
                 onDismissRequest = { showColorPicker.value = false },
                 onColorChanged = { color ->
                 },
-                initialColor = hexToComposeColor( previewUiState.currentColor),
+                initialColor = hexToComposeColor(previewUiState.currentColor),
                 onSave = { color ->
-                    previewViewModel.setCurrentColor(color.toPaletteColor())
+                    previewViewModel.setCurrentColor(color.toPaletteColorString())
                     val paletteCopy = previewUiState.palette.toList().toMutableList()
                     val index = paletteCopy.indexOf(previewUiState.currentColor)
-                    if (index != -1 ) {
-                        paletteCopy[index] = color.toPaletteColor()
+                    if (index != -1) {
+                        paletteCopy[index] = color.toPaletteColorString()
                     }
-                    palleteViewModel.updatePaletteColorsById(previewUiState.paletteID, paletteCopy)
+                    paletteViewModel.updatePaletteColorsById(previewUiState.paletteID, paletteCopy)
                     previewViewModel.setCurrentPalette(paletteCopy)
                     showColorPicker.value = false
                 }
@@ -480,7 +396,8 @@ fun PreviewScreen(previewViewModel: PreviewViewModel = viewModel()) {
             ContrastColorPickerDialog(
                 palette = previewUiState.palette,
                 onDismissRequest = { showContrastModal.value = false },
-                firstColor = previewUiState.currentColor)
+                firstColor = previewUiState.currentColor
+            )
         }
         Column(
             modifier = Modifier
@@ -508,13 +425,13 @@ fun PreviewScreen(previewViewModel: PreviewViewModel = viewModel()) {
                 modifier = Modifier.fillMaxWidth()
             ) {
                 Button(
-                    modifier = Modifier.padding(start=8.dp, top=8.dp, bottom=28.dp),
+                    modifier = Modifier.padding(start = 8.dp, top = 8.dp, bottom = 28.dp),
                     onClick = { showColorPicker.value = true }
                 ) {
                     Text("Edit Color")
                 }
                 Button(
-                    modifier = Modifier.padding(start=8.dp, top=8.dp, bottom=28.dp),
+                    modifier = Modifier.padding(start = 8.dp, top = 8.dp, bottom = 28.dp),
                     onClick = { showContrastModal.value = true }
                 ) {
                     Text("Contrast with Another Color")
